+++ R/utils.R
@@ -1,12 +1,14 @@
+#' @importFrom digest digest
+
 # misc -------------------------------------------------------------------------
+
 #' @noRd
-to_sentence_case <- function(string) {
+to_sentence_case <- function(string)
   paste0(substring(toupper(string), 1L, 1L), substring(tolower(string), 2L))
-}
 
 #' @noRd
 get_next_lowest_factor <-
-  function(x, y) ifelse(x %% y, get_next_lowest_factor(x, y - 1L), y)
+  function(x, y) if (x %% y) get_next_lowest_factor(x, y - 1L) else y
 
 #' @noRd
 #' @importFrom methods as
@@ -23,21 +25,49 @@ pb_head <- function(msg) {
 }
 
 #' @noRd
-truncate_string <- function(x, sl = 20L) {
+truncate_string <- function(x, sl = 20L)
   ifelse(nchar(x) > sl, sprintf("%s\u2026", substr(x, 1L, sl - 1L)), x)
+
+# random sampling --------------------------------------------------------------
+
+#' @noRd
+sample_with_seed <- function(n, size, seed) {
+  if (exists(".Random.seed", 1L)) {
+    oldseed <- get(".Random.seed", 1L)
+    on.exit(assign(".Random.seed", oldseed, 1L))
+  } else {
+    on.exit(rm(".Random.seed", pos = 1L))
+  }
+  args <- list(seed, "default", "default")
+  if (getRversion() >= "3.6.0") args <- c(args, "default")
+  do.call(set.seed, args)
+  sample.int(n, size)
+}
+
+#' @noRd
+gen_seed <- function(x, ...) UseMethod("gen_seed", x)
+
+#' @export
+#' @noRd
+gen_seed.finbif_records_list <- function(x, ...) {
+  hash <- lapply(x, getElement, "hash")
+  hash <- do.call(paste0, hash)
+  hash <- digest::digest(hash)
+  hash <- substr(hash, 1L, 7L)
+  strtoi(hash, 16L)
 }
 
 # errors -----------------------------------------------------------------------
 # modified from https://github.com/reside-ic/defer/blob/master/R/defer.R
+
 #' @noRd
-deferrable_error <- function(message) {
+deferrable_error <- function(message)
   withRestarts({
     calls <- sys.calls()
     call <- calls[[max(length(calls) - 1L, 1L)]]
     stop(error(message, "deferrable_error", call = call, calls = calls))
   },
   continue_deferrable_error = function(...) NULL)
-}
 
 #' @noRd
 defer_errors <- function(expr, handler = stop) {
@@ -59,7 +89,7 @@ defer_errors <- function(expr, handler = stop) {
 }
 
 #' @noRd
-deferred_errors <- function(errors, handler, calls, value = NULL) {
+deferred_errors <- function(errors, handler, calls, value = NULL)
   if (length(errors)) {
     err <- list(errors = errors, value = value)
     class(err) <- c("dfrd_errors", "error", "condition")
@@ -67,14 +97,12 @@ deferred_errors <- function(errors, handler, calls, value = NULL) {
   } else {
     value
   }
-}
 
 #' @noRd
-error <- function(message, class, ...) {
+error <- function(message, class, ...)
   structure(
     list(message = message, ...), class = c(class, "error", "condition")
   )
-}
 
 #' @export
 #' @noRd