+++ tests/testthat/test-finbif_occurrence.R
@@ -11,7 +11,7 @@ vcr::use_cassette(
         expect_s3_class(
           finbif_occurrence(
             species = "Rangifer tarandus fennicus", check_taxa = FALSE,
-            select = "record_id"
+            select = "record_id", sample = TRUE
           ),
           "finbif_occ"
         )
@@ -30,6 +30,23 @@ vcr::use_cassette(
             "Rangifer tarandus fennicus",
             select = c("record_id", "date_start", "lat_wgs84", "lon_wgs84")
           ),
+          "finbif_occ"
+        )
+
+        expect_s3_class(
+          finbif_occurrence(
+            "Pteromys volans", sample = TRUE, n = 1001, cache = FALSE
+          ),
+          "finbif_occ"
+        )
+
+        expect_s3_class(
+          finbif_occurrence("Pteromys volans", sample = TRUE, n = 1001),
+          "finbif_occ"
+        )
+
+        expect_s3_class(
+          finbif_occurrence(select =  "taxon_id", sample = TRUE, n = 3001),
           "finbif_occ"
         )