+++ NEWS.md
@@ -1,3 +1,13 @@
+# finbif 0.2.0
+
+### NEW FEATURES
+
+  * Add capacity to request a random sample of FinBIF records.
+
+### MINOR IMPROVEMENTS
+
+  * Add more content to vignettes.
+
 # finbif 0.1.0
 
   * Initial release.